proj1/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
azure-core==1.28.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.2.0
Click==7.0
cloud-init==23.1.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dataclasses-json==0.5.14
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
frozenlist==1.4.0
greenlet==2.0.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
langchain==0.0.174
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
MarkupSafe==1.1.0
marshmallow==3.20.1
more-itertools==4.2.0
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.10.4
numexpr==2.8.4
numpy==1.24.4
oauthlib==3.1.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pexpect==4.6.0
protobuf==3.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
pydantic==1.10.12
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
pytz==2019.3
PyYAML==6.0.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.4
SQLAlchemy==2.0.19
ssh-import-id==5.10
systemd-python==234
tenacity==8.2.2
tqdm==4.65.0
Twisted==18.9.0
typing-extensions==4.7.1
typing-inspect==0.9.0
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
yarl==1.9.2
zipp==1.0.0
zope.interface==4.7.1
